api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
google-cloud-speech==2.23.0
google-generativeai==0.3.2
python-dotenv==1.0.1
pydantic==2.6.1
